Renamed 'nerman-feeds' to 'feeds'.
