--- commands/slash/feeds/add/farcaster.js
@@ -1,10 +1,0 @@
-const add = require('./add');
-module.exports = {
-   subCommand: 'nerman-feeds.add.farcaster',
-   /**
-    * @param {CommandInteraction} interaction
-    */
-   async execute(interaction) {
-      add.execute(interaction);
-   },
-};
