--- commands/slash/feeds/add/federation.js
@@ -1,10 +1,0 @@
-const add = require('./add');
-module.exports = {
-   subCommand: 'nerman-feeds.add.federation',
-   /**
-    * @param {CommandInteraction} interaction
-    */
-   async execute(interaction) {
-      add.execute(interaction);
-   },
-};
