--- commands/slash/feeds/add/lil-nouns.js
@@ -1,10 +1,0 @@
-const add = require('./add');
-module.exports = {
-   subCommand: 'nerman-feeds.add.lil-nouns',
-   /**
-    * @param {CommandInteraction} interaction
-    */
-   async execute(interaction) {
-      add.execute(interaction);
-   },
-};
