--- commands/slash/feeds/add/nouns-nymz.js
@@ -1,10 +1,0 @@
-const add = require('./add');
-module.exports = {
-   subCommand: 'nerman-feeds.add.nouns-nymz',
-   /**
-    * @param {CommandInteraction} interaction
-    */
-   async execute(interaction) {
-      add.execute(interaction);
-   },
-};
